+++ __tests__/setup/pageObjects/getConsoleLogs.js
@@ -1,0 +1,3 @@
+export default async function getConsoleLogs(driver) {
+  return await driver.executeScript(() => window.__console__);
+}