+++ samples/19.a.single-sign-on-for-enterprise-apps/app/src/microsoftGraphProfile/Composer.js
@@ -10,12 +10,7 @@ import MicrosoftGraphProfileContext from './Context';
 import OAuthComposer from '../oauth/Composer';
 
 // Composer is a React component with a React context
-const MicrosoftGraphProfileComposer = ({
-  accessToken,
-  children,
-  onSignIn,
-  onSignOut
-}) => {
+const MicrosoftGraphProfileComposer = ({ accessToken, children, onSignIn, onSignOut }) => {
   const [avatarURL, setAvatarURL] = useState('');
   const [name, setName] = useState('');
 
@@ -31,24 +26,18 @@ const MicrosoftGraphProfileComposer = ({
   }, [accessToken]);
 
   // Build a new React context object if anything changed.
-  const context = useMemo(() => ({
-    avatarURL,
-    name,
-    onSignIn,
-    onSignOut
-  }), [
-    avatarURL,
-    name,
-    onSignIn,
-    onSignOut
-  ]);
-
-  return (
-    <MicrosoftGraphProfileContext.Provider value={ context }>
-      { children }
-    </MicrosoftGraphProfileContext.Provider>
+  const context = useMemo(
+    () => ({
+      avatarURL,
+      name,
+      onSignIn,
+      onSignOut
+    }),
+    [avatarURL, name, onSignIn, onSignOut]
   );
-}
+
+  return <MicrosoftGraphProfileContext.Provider value={context}>{children}</MicrosoftGraphProfileContext.Provider>;
+};
 
 MicrosoftGraphProfileComposer.defaultProps = {
   accessToken: '',
@@ -61,7 +50,7 @@ MicrosoftGraphProfileComposer.propTypes = {
   accessToken: PropTypes.string,
   children: PropTypes.any,
   onSignIn: PropTypes.func,
-  onSignOut: PropTypes.func,
+  onSignOut: PropTypes.func
 };
 
 // Hoist the functionality from generic OAuth composer to GitHub composer.
@@ -70,23 +59,17 @@ MicrosoftGraphProfileComposer.propTypes = {
 const ComposedMicrosoftGraphProfileComposer = compose(
   connectSignInButton(({ onClick }) => ({ onSignIn: onClick })),
   connectSignOutButton(({ onClick }) => ({ onSignOut: onClick }))
-)(MicrosoftGraphProfileComposer)
+)(MicrosoftGraphProfileComposer);
 
 // This is the exported React component, which provides basic UI-less functionality to its descendants.
-const ConnectedMicrosoftGraphProfileComposer = ({
-  accessToken,
-  children,
-  oauthAuthorizeURL,
-  onAccessTokenChange
-}) =>
+const ConnectedMicrosoftGraphProfileComposer = ({ accessToken, children, oauthAuthorizeURL, onAccessTokenChange }) => (
   <OAuthComposer
-    accessToken={ accessToken }
-    oauthAuthorizeURL={ oauthAuthorizeURL }
-    onAccessTokenChange={ onAccessTokenChange }
+    accessToken={accessToken}
+    oauthAuthorizeURL={oauthAuthorizeURL}
+    onAccessTokenChange={onAccessTokenChange}
   >
-    <ComposedMicrosoftGraphProfileComposer accessToken={ accessToken }>
-      { children }
-    </ComposedMicrosoftGraphProfileComposer>
+    <ComposedMicrosoftGraphProfileComposer accessToken={accessToken}>{children}</ComposedMicrosoftGraphProfileComposer>
   </OAuthComposer>
+);
 
-export default ConnectedMicrosoftGraphProfileComposer
+export default ConnectedMicrosoftGraphProfileComposer;