+++ samples/19.a.single-sign-on-for-enterprise-apps/app/src/utils/hocContext.js
@@ -2,14 +2,9 @@ import { createElement } from 'react';
 
 // Helper function to hoist a component as HOC with corresponding props from a specified React context.
 export default (context, selector = state => state) => component => props =>
-  createElement(
-    context.Consumer,
-    {},
-    state => createElement(
-      component,
-      {
-        ...props,
-        ...selector(state)
-      }
-    )
-  )
+  createElement(context.Consumer, {}, state =>
+    createElement(component, {
+      ...props,
+      ...selector(state)
+    })
+  );