+++ samples/19.a.single-sign-on-for-enterprise-apps/bot/src/createBotAdapter.js
@@ -1,9 +1,6 @@
 const { BotFrameworkAdapter } = require('botbuilder');
 
-const {
-  MICROSOFT_APP_ID,
-  MICROSOFT_APP_PASSWORD
-} = process.env;
+const { MICROSOFT_APP_ID, MICROSOFT_APP_PASSWORD } = process.env;
 
 module.exports = function createBotAdapter() {
   const adapter = new BotFrameworkAdapter({