+++ samples/21.customization-plain-ui/src/App.js
@@ -1,8 +1,8 @@
-import { Components } from "botframework-webchat-component";
-import { createDirectLine } from "botframework-webchat";
-import React from "react";
+import { Components } from 'botframework-webchat-component';
+import { createDirectLine } from 'botframework-webchat';
+import React from 'react';
 
-import PlainWebChat from "./PlainWebChat";
+import PlainWebChat from './PlainWebChat';
 
 // In this demo, we are using Direct Line token from MockBot.
 // To talk to your bot, you should use the token exchanged using your Direct Line secret.
@@ -10,10 +10,7 @@ import PlainWebChat from "./PlainWebChat";
 // https://docs.microsoft.com/en-us/azure/bot-service/rest-api/bot-framework-rest-direct-line-3-0-authentication
 
 async function getDirectLineToken() {
-  const res = await fetch(
-    "https://webchat-mockbot.azurewebsites.net/directline/token",
-    { method: "POST" }
-  );
+  const res = await fetch('https://webchat-mockbot.azurewebsites.net/directline/token', { method: 'POST' });
   const { token } = await res.json();
 
   return token;
@@ -24,9 +21,7 @@ export default () => {
 
   if (!directLine) {
     // We will load DirectLineJS asynchronously on first render.
-    getDirectLineToken().then(token =>
-      setDirectLine(createDirectLine({ token }))
-    );
+    getDirectLineToken().then(token => setDirectLine(createDirectLine({ token })));
   }
 
   return (
@@ -34,8 +29,8 @@ export default () => {
     <React.Fragment>
       <h1>Web Chat with plain UI</h1>
       <p>
-        This sample shows how to use Web Chat without any of its canned UI
-        component. There are few conversation you can try out.
+        This sample shows how to use Web Chat without any of its canned UI component. There are few conversation you can
+        try out.
       </p>
       <ol>
         <li>
@@ -49,15 +44,11 @@ export default () => {
         </li>
       </ol>
       <p>
-        For the{" "}
+        For the{' '}
         <a href="https://github.com/microsoft/BotFramework-WebChat/tree/master/samples/21.customization-plain-ui/">
           source code of this demo
         </a>
-        , please visit{" "}
-        <a href="https://github.com/microsoft/BotFramework-WebChat/">
-          our GitHub repository
-        </a>
-        .
+        , please visit <a href="https://github.com/microsoft/BotFramework-WebChat/">our GitHub repository</a>.
       </p>
       <hr />
       {!!directLine && (